LPS-82164 Extract custom action logic to helper classes on forms headless
