+++ modules/apps/headless-apio/forms/forms-apio-impl/src/main/java/com/liferay/forms/apio/internal/architect/resource/FormInstanceRecordNestedCollectionResource.java
@@ -18,8 +18,6 @@ import static com.liferay.forms.apio.internal.util.FormInstanceRecordResourceUti
 import static com.liferay.forms.apio.internal.util.FormValuesUtil.getDDMFormValues;
 import static com.liferay.forms.apio.internal.util.LocalizedValueUtil.getLocalizedString;
 
-import com.google.gson.Gson;
-
 import com.liferay.apio.architect.functional.Try;
 import com.liferay.apio.architect.pagination.PageItems;
 import com.liferay.apio.architect.pagination.Pagination;
@@ -27,7 +25,6 @@ import com.liferay.apio.architect.representor.Representor;
 import com.liferay.apio.architect.resource.NestedCollectionResource;
 import com.liferay.apio.architect.routes.ItemRoutes;
 import com.liferay.apio.architect.routes.NestedCollectionRoutes;
-import com.liferay.document.library.kernel.service.DLAppService;
 import com.liferay.dynamic.data.mapping.model.DDMForm;
 import com.liferay.dynamic.data.mapping.model.DDMFormField;
 import com.liferay.dynamic.data.mapping.model.DDMFormInstance;
@@ -35,8 +32,6 @@ import com.liferay.dynamic.data.mapping.model.DDMFormInstanceRecord;
 import com.liferay.dynamic.data.mapping.model.DDMFormInstanceRecordVersion;
 import com.liferay.dynamic.data.mapping.model.DDMFormInstanceRecordVersionModel;
 import com.liferay.dynamic.data.mapping.model.DDMStructure;
-import com.liferay.dynamic.data.mapping.model.UnlocalizedValue;
-import com.liferay.dynamic.data.mapping.model.Value;
 import com.liferay.dynamic.data.mapping.service.DDMFormInstanceRecordService;
 import com.liferay.dynamic.data.mapping.service.DDMFormInstanceService;
 import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
@@ -45,6 +40,7 @@ import com.liferay.forms.apio.architect.identifier.FormInstanceIdentifier;
 import com.liferay.forms.apio.architect.identifier.FormInstanceRecordIdentifier;
 import com.liferay.forms.apio.internal.architect.form.FormInstanceRecordForm;
 import com.liferay.forms.apio.internal.architect.locale.AcceptLocale;
+import com.liferay.forms.apio.internal.helper.UploadFileHelper;
 import com.liferay.forms.apio.internal.model.FileEntryValue;
 import com.liferay.forms.apio.internal.model.ServiceContextWrapper;
 import com.liferay.forms.apio.internal.util.FormInstanceRecordResourceUtil;
@@ -53,11 +49,7 @@ import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
 
-import java.util.Collection;
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.stream.Stream;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -176,42 +168,11 @@ public class FormInstanceRecordNestedCollectionResource
 
 		List<DDMFormField> ddmFormFields = ddmForm.getDDMFormFields();
 
-		_linkFiles(ddmFormFields, ddmFormValues.getDDMFormFieldValues());
+		_uploadFileHelper.linkFiles(
+			ddmFormFields, ddmFormValues.getDDMFormFieldValues());
 
 		return _ddmFormInstanceRecordService.addFormInstanceRecord(
 			groupId, formInstanceId, ddmFormValues, serviceContext);
-	}
-
-	private Long _extractFileEntryId(DDMFormFieldValue ddmFormFieldValue) {
-		return Try.fromFallible(
-			() -> ddmFormFieldValue.getValue()
-		).map(
-			Value::getValues
-		).map(
-			Map::values
-		).map(
-			Collection::stream
-		).mapOptional(
-			Stream::findFirst
-		).map(
-			fileEntryUrl -> fileEntryUrl.substring(
-				fileEntryUrl.lastIndexOf("/") + 1)
-		).map(
-			Long::valueOf
-		).orElse(
-			null
-		);
-	}
-
-	private Optional<DDMFormFieldValue> _findField(
-		DDMFormField formField, List<DDMFormFieldValue> ddmFormFieldValues) {
-
-		Stream<DDMFormFieldValue> ddmFormFieldValuesStream =
-			ddmFormFieldValues.stream();
-
-		return ddmFormFieldValuesStream.filter(
-			value -> value.getName().equals(formField.getName())
-		).findFirst();
 	}
 
 	private List<DDMFormFieldValue> _getFieldValues(
@@ -240,42 +201,6 @@ public class FormInstanceRecordNestedCollectionResource
 			formInstanceId);
 
 		return new PageItems<>(ddmFormInstanceRecords, count);
-	}
-
-	private void _linkFiles(
-		List<DDMFormField> ddmFormFields,
-		List<DDMFormFieldValue> ddmFormFieldValues) {
-
-		Stream<DDMFormField> ddmFormFieldsStream = ddmFormFields.stream();
-
-		ddmFormFieldsStream.filter(
-			formField -> formField.getType().equals("document_library")
-		).map(
-			field -> _findField(field, ddmFormFieldValues)
-		).forEach(
-			optional -> optional.ifPresent(this::_setFileEntryAsFormFieldValue)
-		);
-	}
-
-	private void _setFileEntryAsFormFieldValue(
-		DDMFormFieldValue ddmFormFieldValue) {
-
-		Gson gson = new Gson();
-
-		Try.fromFallible(
-			() -> _extractFileEntryId(ddmFormFieldValue)
-		).map(
-			_dlAppService::getFileEntry
-		).map(
-			fileEntry -> new FileEntryValue(
-				fileEntry.getGroupId(), fileEntry.getUuid())
-		).map(
-			gson::toJson
-		).map(
-			UnlocalizedValue::new
-		).ifSuccess(
-			ddmFormFieldValue::setValue
-		);
 	}
 
 	private DDMFormInstanceRecord _updateFormInstanceRecord(
@@ -314,6 +239,6 @@ public class FormInstanceRecordNestedCollectionResource
 	private DDMFormInstanceService _ddmFormInstanceService;
 
 	@Reference
-	private DLAppService _dlAppService;
+	private UploadFileHelper _uploadFileHelper;
 
 }
